requirements.txt
addict==2.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.0
Cython==0.29.33
editdistance==0.6.2
filelock==3.10.0
idna==3.4
Jinja2==3.1.2
lit==15.0.7
MarkupSafe==2.1.2
mmcv==1.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.6.0.66
packaging==23.0
Pillow==9.4.0
Polygon3==3.0.9.1
prefetch-generator==1.0.3
pyclipper==1.3.0.post4
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
yacs==0.1.8
yapf==0.32.0
